+++ MangoPay.SDK.Tests/ApiPayInsTest.cs
@@ -118,8 +118,9 @@ namespace MangoPay.SDK.Tests
                         FirstName = user.FirstName,
                         LastName = user.LastName
                     },
-                    "test"
+                    "test",
                     // CultureCode.FR
+                    shippingPreference: ShippingPreference.SET_PROVIDED_ADDRESS
                 );
                 
                 var payIn = await this.Api.PayIns.CreatePayPalWebAsync(payInPost);
@@ -253,9 +254,9 @@ namespace MangoPay.SDK.Tests
         {
             try
             {
-                var user = await this.GetJohn();
-                var payIn = await this.GetNewPayInMbwayWeb();
-                var fetched = await this.Api.PayIns.GetMbwayAsync(payIn.Id);
+                var user = await GetJohn();
+                var payIn = await GetNewPayInMbwayWeb();
+                var fetched = await Api.PayIns.GetMbwayAsync(payIn.Id);
 
                 Assert.IsTrue(payIn.Id.Length > 0);
                 Assert.AreEqual(PayInPaymentType.MBWAY, payIn.PaymentType);
@@ -270,6 +271,122 @@ namespace MangoPay.SDK.Tests
                 
                 Assert.AreEqual(payIn.Id, fetched.Id);
 
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+
+        [Test]
+        public async Task Test_PayIns_Create_MultibancoWeb()
+        {
+            try
+            {
+                var user = await GetJohn();
+                var payIn = await GetNewPayInMultibancoWeb();
+                var fetched = await Api.PayIns.GetMultibancoAsync(payIn.Id);
+
+                Assert.IsTrue(payIn.Id.Length > 0);
+                Assert.AreEqual(PayInPaymentType.MULTIBANCO, payIn.PaymentType);
+                Assert.AreEqual(PayInExecutionType.WEB, payIn.ExecutionType);
+                Assert.IsTrue(payIn.DebitedFunds is Money);
+                Assert.IsTrue(payIn.CreditedFunds is Money);
+                Assert.IsTrue(payIn.Fees is Money);
+                Assert.AreEqual(user.Id, payIn.AuthorId);
+                Assert.AreEqual(TransactionStatus.CREATED, payIn.Status);
+                Assert.AreEqual(TransactionType.PAYIN, payIn.Type);
+                Assert.AreEqual(TransactionNature.REGULAR, payIn.Nature);
+                
+                Assert.AreEqual(payIn.Id, fetched.Id);
+                Assert.IsNotNull(fetched.Id);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+        
+        [Test]
+        public async Task Test_PayIns_Create_SatispayWeb()
+        {
+            try
+            {
+                var user = await GetJohn();
+                var payIn = await GetNewPayInSatispayWeb();
+                var fetched = await Api.PayIns.GetSatispayAsync(payIn.Id);
+
+                Assert.IsTrue(payIn.Id.Length > 0);
+                Assert.AreEqual(PayInPaymentType.SATISPAY, payIn.PaymentType);
+                Assert.AreEqual(PayInExecutionType.WEB, payIn.ExecutionType);
+                Assert.IsTrue(payIn.DebitedFunds is Money);
+                Assert.IsTrue(payIn.CreditedFunds is Money);
+                Assert.IsTrue(payIn.Fees is Money);
+                Assert.AreEqual(user.Id, payIn.AuthorId);
+                Assert.AreEqual(TransactionStatus.CREATED, payIn.Status);
+                Assert.AreEqual(TransactionType.PAYIN, payIn.Type);
+                Assert.AreEqual(TransactionNature.REGULAR, payIn.Nature);
+                
+                Assert.AreEqual(payIn.Id, fetched.Id);
+                Assert.IsNotNull(fetched.Id);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+        
+        [Test]
+        public async Task Test_PayIns_Create_BlikWeb()
+        {
+            try
+            {
+                var user = await GetJohn();
+                var payIn = await GetNewPayInBlikWeb();
+                var fetched = await Api.PayIns.GetBlikAsync(payIn.Id);
+
+                Assert.IsTrue(payIn.Id.Length > 0);
+                Assert.AreEqual(PayInPaymentType.BLIK, payIn.PaymentType);
+                Assert.AreEqual(PayInExecutionType.WEB, payIn.ExecutionType);
+                Assert.IsTrue(payIn.DebitedFunds is Money);
+                Assert.IsTrue(payIn.CreditedFunds is Money);
+                Assert.IsTrue(payIn.Fees is Money);
+                Assert.AreEqual(user.Id, payIn.AuthorId);
+                Assert.AreEqual(TransactionStatus.CREATED, payIn.Status);
+                Assert.AreEqual(TransactionType.PAYIN, payIn.Type);
+                Assert.AreEqual(TransactionNature.REGULAR, payIn.Nature);
+                
+                Assert.AreEqual(payIn.Id, fetched.Id);
+                Assert.IsNotNull(fetched.Id);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+        
+        [Test]
+        public async Task Test_PayIns_Create_KlarnaWeb()
+        {
+            try
+            {
+                var user = await GetJohn();
+                var payIn = await GetNewPayInKlarnaWeb();
+                var fetched = await Api.PayIns.GetKlarnaAsync(payIn.Id);
+
+                Assert.IsTrue(payIn.Id.Length > 0);
+                Assert.AreEqual(PayInPaymentType.KLARNA, payIn.PaymentType);
+                Assert.AreEqual(PayInExecutionType.WEB, payIn.ExecutionType);
+                Assert.IsTrue(payIn.DebitedFunds is Money);
+                Assert.IsTrue(payIn.CreditedFunds is Money);
+                Assert.IsTrue(payIn.Fees is Money);
+                Assert.AreEqual(user.Id, payIn.AuthorId);
+                Assert.AreEqual(TransactionStatus.CREATED, payIn.Status);
+                Assert.AreEqual(TransactionType.PAYIN, payIn.Type);
+                Assert.AreEqual(TransactionNature.REGULAR, payIn.Nature);
+                
+                Assert.AreEqual(payIn.Id, fetched.Id);
+                Assert.IsNotNull(fetched.Id);
             }
             catch (Exception ex)
             {
@@ -648,6 +765,30 @@ namespace MangoPay.SDK.Tests
             Assert.AreEqual(getPayIn.AuthorId, googlePayIn.AuthorId);
             Assert.AreEqual(getPayIn.PaymentType, PayInPaymentType.GOOGLEPAY);
             Assert.AreEqual(getPayIn.Status, TransactionStatus.SUCCEEDED);
+        }
+
+        [Test]
+        public async Task Test_PayIns_Create_GooglePayDirect()
+        {
+            try
+            {
+                var user = await GetJohn();
+                var payIn = await GetNewPayInGooglePayDirect();
+
+                Assert.IsTrue(payIn.Id.Length > 0);
+                Assert.AreEqual(PayInPaymentType.GOOGLE_PAY, payIn.PaymentType);
+                Assert.AreEqual(PayInExecutionType.DIRECT, payIn.ExecutionType);
+                Assert.IsTrue(payIn.DebitedFunds is Money);
+                Assert.IsTrue(payIn.CreditedFunds is Money);
+                Assert.IsTrue(payIn.Fees is Money);
+                Assert.AreEqual(user.Id, payIn.AuthorId);
+                Assert.AreEqual(TransactionType.PAYIN, payIn.Type);
+                Assert.AreEqual(TransactionNature.REGULAR, payIn.Nature);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
         }
 
         [Test]