+++ MangoPay.SDK/Core/Enumerations/PayInPaymentType.cs
@@ -27,6 +27,9 @@ namespace MangoPay.SDK.Core.Enumerations
 
         /// <summary> GooglePay payment type </summary>
         GOOGLEPAY,
+        
+        /// <summary> GooglePay V2 payment type </summary>
+        GOOGLE_PAY,
 
         /// <summary>
         /// Payconiq payment type 
@@ -35,5 +38,13 @@ namespace MangoPay.SDK.Core.Enumerations
         
         /// <summary> Mbway payment type </summary>
         MBWAY,
+        /// <summary> Multibanco payment type </summary>
+        MULTIBANCO,
+        /// <summary> Satispay payment type </summary>
+        SATISPAY,
+        /// <summary> Blik payment type </summary>
+        BLIK,
+        /// <summary> Klarna payment type </summary>
+        KLARNA,
     }
 }